avoid duplication when implementing delete
