+++ src/lib.rs
@@ -15,7 +15,7 @@ pub use traits::{Bind, BindableQuery, Create, Crud, Delete, Schema, Update};
 /// Prelude including all the necessary traits for convenience
 pub mod prelude {
     pub use super::store::Store;
-    pub use super::traits::{Bind, BindableQuery, Schema};
+    pub use super::traits::{Bind, BindableQuery, RowsAffected, Schema};
     pub use super::traits::{Create, Crud, Delete, Update};
     pub use miniorm_macros::Entity;
 }
