+++ src/traits/mod.rs
@@ -1,7 +1,9 @@
 mod bind;
 mod crud;
 mod schema;
+mod sqlx_ext;
 
-pub use bind::{Bind, BindableQuery};
+pub use bind::Bind;
 pub use crud::{Create, Crud, Delete, Update};
 pub use schema::Schema;
+pub use sqlx_ext::{BindableQuery, RowsAffected};
